+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### master
 
+- rename `selectAs` to `selectAsJson`
 - add `primaryKey` column marker to create correct group-by statements for `selectAsJsonAgg`
 - add `whereSql` and the `sql` tagged template function